RegressaoLinearMultipla/SeguroMedico/requirements.txt
numpy==1.23.5
pandas==1.5.3
Pillow==9.5.0
requests==2.28.1
streamlit==1.22.0
scikit-learn
